spoonacular/requirements.txt
fastapi==0.109.2
httpx==0.26.0
pydantic==2.6.1
python-dotenv==1.0.1
uvicorn==0.27.0
